\docType{package}
\name{grex-package}
\alias{grex}
\alias{grex-package}
\title{EXecuting simulations with parameters based on a GRid}
\description{
This package simplifies the setup of simulation experiments with
more than one input parameter, with support for parallelization
}
\details{
\tabular{ll}{
  Package: \tab grex\cr
  Type: \tab Package\cr
  Version: \tab 0.0.2\cr
  Date: \tab 2014-02-01\cr
  License: \tab GPL-3\cr
} ~~
}
\examples{
# ~~ simple examples of the most important functions ~~
}
\author{
Kirill Müller: \email{kirill.mueller@ivt.baug.ethz.ch>}

Submit bug reports to: \url{https://github.com/krlmlr/grex/issues}
}
\keyword{experiments,}
\keyword{package,}
\keyword{parallelization}
\keyword{simulation,}

