% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grex-datalist.R
\docType{data}
\name{gtexv6}
\alias{gtexv6}
\title{Ensembl IDs from GTEx V6 Gene Read Count Data}
\format{A character vector with 56,318 Ensembl IDs.}
\source{
\url{https://www.gtexportal.org}
}
\usage{
gtexv6
}
\description{
A dataset containing the Ensembl IDs from GTEx (V6)
gene read count data.
}
\keyword{datasets}

