% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grex-datalist.R
\docType{data}
\name{gtexv6p}
\alias{gtexv6p}
\title{Ensembl IDs from GTEx V6p Gene Read Count Data}
\format{
A character vector with 56,238 Ensembl IDs.
}
\source{
\url{https://www.gtexportal.org}
}
\usage{
gtexv6p
}
\description{
A dataset containing the Ensembl IDs from GTEx (V6p)
gene read count data.
}
\keyword{datasets}
