% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grex-datalist.R
\docType{data}
\name{gtexv7}
\alias{gtexv7}
\title{Ensembl IDs from GTEx V7 Gene Read Count Data}
\format{A character vector with 56,202 Ensembl IDs.}
\source{
\url{https://www.gtexportal.org}
}
\usage{
gtexv7
}
\description{
A dataset containing the Ensembl IDs from GTEx (V7)
gene read count data.
}
\keyword{datasets}
