% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grex-package.R
\docType{package}
\name{grex-package}
\alias{grex-package}
\title{Gene ID Mapping for Genotype-Tissue Expression (GTEx) Data}
\description{
Gene ID Mapping for Genotype-Tissue Expression (GTEx) Data
}
\details{
Opened the vignette via \code{vignette("grex")}.

\tabular{ll}{ Package: \tab grex\cr
Type: \tab Package\cr
License: \tab GPL-3\cr }
}
\author{
Nan Xiao <\email{me@nanx.me}>,
        Gao Wang <\email{gaow@uchicago.edu}>,
        Lei Sun <\email{sunl@uchicago.edu}>
}

