% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grex-package.R
\docType{package}
\name{grex-package}
\alias{grex}
\alias{grex-package}
\title{grex: Gene ID Mapping for Genotype-Tissue Expression (GTEx) Data}
\description{
Convert 'Ensembl' gene identifiers from Genotype-Tissue
    Expression (GTEx) data to identifiers in other annotation systems,
    including 'Entrez', 'HGNC', and 'UniProt'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nanx.me/grex/}
  \item \url{https://github.com/road2stat/grex}
  \item Report bugs at \url{https://github.com/road2stat/grex/issues}
}

}
\author{
\strong{Maintainer}: Nan Xiao \email{me@nanx.me} (0000-0002-0250-5673)

Authors:
\itemize{
  \item Gao Wang \email{gaow@uchicago.edu}
  \item Lei Sun \email{sunl@uchicago.edu}
}

}
\keyword{internal}
